pyproject.toml
[project]
name = "project-sync-service"
version = "0.1.0"
description = "Sync service for pulling project data from FileMaker (UCPPC) into PostgreSQL (business_services_db)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fmrest>=1.1.2",
    "psycopg[binary]>=3.1",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "python-dateutil>=2.8",
]

[project.scripts]
project-sync = "project_sync_service.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/project_sync_service"]
